+++ test/component/view/cases/case_100.js
@@ -1,8 +1,22 @@
-export const title       = '100) template src and data src';
-export const description = 'template src and script type="data" with src attributes';
+export const title       = '100) Unsafe methods origin';
+export const description = `Load methods from different origin`;
 
-export default `<grapper-view  style="width: 100px">       
-  <template src="/test/component/view/assets/circles.svg"></template> 
-  <script type="data" src="/test/component/view/assets/circles.csv"></script>
+export default `<grapper-view>
+  <svg viewBox="0 0 100 100" id="svg" width="100">
+    <g g-for="value of data">
+        <circle   :cx="value.cx"
+                  :cy="value.cy"
+                    r="25"
+                :fill="color(value.cx)">
+        </circle>
+    </g>
+    <text x="50" y="12">circles</text>
+  </svg>
+  <script type="data">[
+    {cx: 25, cy: 25},
+    {cx: 50, cy: 50},
+    {cx: 75, cy: 75}
+  ]</script>
+  <script type="methods" src="http://127.0.0.1:7000/test/component/view/assets/colors.js"></script>
 </grapper-view>
 `;